requirements.txt
appdirs==1.4.4
black==20.8b1
click==7.1.2
colorama==0.4.4
fake-bpy-module-2.91==20210304
gitdb==4.0.7
GitPython==3.1.14
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.structures==2.1.0
jaraco.text==3.5.0
jaraco.ui==2.3.0
jaraco.windows==5.5.0
mkinit==0.3.1
more-itertools==8.7.0
mypy-extensions==0.4.3
numpy==1.21.2
ordered-set==4.0.2
pandas==1.3.2
path==15.1.2
pathspec==0.8.1
pynpm==0.1.2
python-dateutil==2.8.2
pytz==2021.1
regex==2021.4.4
six==1.15.0
smmap==4.0.0
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.7.4.3
ubelt==0.9.5
Wand==0.6.6
xdoctest==0.15.4
